イベント発生の種類の追加、DB処理の修正
プレイヤーが宝くじを購入しようとした際のイベントと何らかのエラーでこのプラグインが返金処理をしたときに発生されるイベントの追加
プレイヤーの買ったチケット枚数が正しくカウントされなかった問題の修正

+++ src/main/java/com/example/minicash/lottery/database/ActiveDatabase.java
@@ -165,7 +165,7 @@ public class ActiveDatabase {
             try (Connection connection = hikariDataSource.getConnection();
                  PreparedStatement stmt = connection.prepareStatement(updateSql)) {
 
-                stmt.setDouble(1, amount);
+                stmt.setInt(1, amount);
                 stmt.setString(2, sessionID);
 
                 int updatedRows = stmt.executeUpdate();
